Added ensemble analysis and conformational inference utilities.
Add tests for residue states and conditional logits functionality

- Implemented unit tests for residue states functionality in `test_residue_states.py`, covering various scenarios including empty ensembles and single-frame ensembles.
- Added integration tests for conditional logits in `test_conditional_logits.py`, ensuring correct behavior with mock model parameters and various input conditions.
- Updated sampling tests in `test_sample.py` to use int8 for initial sequences.
- Removed assertions related to b-factors in `test_io.py` and `test_data_structures.py` as they are no longer relevant.
- Introduced new tests for entropy utility functions in `test_entropy.py`, covering von Neumann entropy, MLE entropy, and posterior entropy calculations.
- Cleaned up unused b_factors in various test files to streamline the codebase.

+++ tests/sampling/test_sample.py
@@ -119,7 +119,7 @@ def test_make_sample_sequences(mock_model_parameters):
   residue_indices = jnp.arange(L)
   chain_indices = jnp.zeros((L,))
   key, sequence_key = jax.random.split(key)
-  initial_sequence = jax.random.randint(sequence_key, (L,), 0, 21, dtype=jnp.int32)
+  initial_sequence = jax.random.randint(sequence_key, (L,), 0, 21, dtype=jnp.int8)
   key, sample_key = jax.random.split(key)
 
   # Run sampling
